refactoring workshop component
